requirements.txt
tensorflow>=2.12.0,<2.21.0
streamlit
numpy
pillow